feat: implement initial GNN traffic dashboard with real-time monitoring, configuration, and UI components.
